Fixed $sp off by one word bug; see offset math description
Positive offset math is from 0 inclusive to  4 exclusive; however when using a negative offset, as when moving the $sp to lower memory addresses, the offset is 0 exclusive to 4 inclusive; the word in memory you are reading/writing to is ABOVE at higher address instead of BELOW at lower address so the offset back up is +4 more than if you were going down.  Not doing this corrupted the stack pointer value or ruined the return value causing exceptions.